Modify - Review Service 로직 세분화 및 리뷰 수정, 리뷰 설문 로직 변경
- ReviewService 비즈니스 로직 세분화
-- 세분화된 비즈니스 로직을 사용하여 Application Service Layer 에서 사용

- 리뷰 설문 추가 시 테마 분석에 반영되도록 변경

- 리뷰 수정 시 테마 평점 값 수정되도록 변경

- 리뷰 설문 수정 주석 처리
-- 사용하지 않을 경우 삭제할 예정

* 변경된 부분 모두 Test 수정 및 새로 작성 완료
* 변경된 부분 문서 반영 완료
* 프로젝트 build 완료

close #79 #86

+++ src/main/java/bbangduck/bd/bbangduck/domain/member/dto/controller/response/MemberPlayInclinationResponseDto.java
@@ -21,7 +21,7 @@ public class MemberPlayInclinationResponseDto {
 
     private MemberPlayInclinationGenreSimpleInfoResponseDto genre;
 
-    private Integer playCount;
+    private Long playCount;
 
     public static MemberPlayInclinationResponseDto convert(MemberPlayInclination memberPlayInclination) {
         return MemberPlayInclinationResponseDto.builder()
